Deprecate [cocoapods] stats badges; split out tests
Resolve #2651.
